+++ CHANGELOG.md
@@ -36,6 +36,8 @@ with the `YY.MINOR.MICRO` scheme.
 
 - Clusterwide SQL supports procedure calling.
 
+- Clusterwide SQL supports procedure renaming.
+
 -->
 
 ## [24.1.1] - 2024-02-09